Update form-field.html
According to [H:44 Using label elements to associate text labels with form controls](https://www.w3.org/TR/WCAG20-TECHS/H44.html)
> Check that there is a label element that identifies the purpose of the control **before** the input, textarea, or select element.

If I manually change the order in my local copy of the library, the component still functions as expected. So, why was/is the choice made to have the `label` after the `input` in the DOM when that goes against the WCAG?

If you switch their positions, you likely no longer need the `aira-owns` attribute to make JAWS & NVDA happy 😉

Thanks!